+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 * core/json: KClass.createFrom will report missing parameters in a nicer way
 * server: introduced FormDataRenderer
+* core: conversion of enum values from string is now case-insensitive
 
 # 1.6.1
 * jdbc: and() introduced in addition to or() for more convenient composition of where expressions