+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Unreleased
+# 1.6.2
 * core: bugfix: better check for paths being inside of assets directory
 * core/json: KClass.createFrom will report missing parameters in a nicer way
 * server: introduced FormDataRenderer